showroom/usecase-000/requirements.txt
openai>=1.41.0
python-dotenv>=1.0.0
pylint>=3.0.0